Add files via upload
Disclaimer
This tool is provided for educational and demonstrative purposes only. Users are responsible for ensuring the security and legality of their actions when using this software.

Contributions
Contributions are welcome! Feel free to open issues or pull requests to suggest improvements or report bugs.

License
This project is licensed under the MIT License.